sentinel-hub-app/requirements.txt
firebase_admin
flask
sentinelhub
numpy
python-dotenv
requests
pyjwt
werkzeug